rename Messenger to MessengerBatch
