+++ packages/messaging-api-messenger/src/index.js
@@ -1,9 +1,11 @@
 /* @flow */
 
 import Messenger from './Messenger';
+import MessengerBatch from './MessengerBatch';
 import MessengerClient from './MessengerClient';
 
 module.exports = {
   Messenger,
+  MessengerBatch,
   MessengerClient,
 };
